[CHORE] Remove Useless Local Property
