+++ app/build.gradle.kts
@@ -12,7 +12,6 @@ plugins {
 
 val properties = Properties()
 properties.load(project.rootProject.file("local.properties").inputStream())
-
 android {
     compileSdk = DefaultConfig.COMPILE_SDK
 
@@ -22,14 +21,8 @@ android {
         targetSdk = DefaultConfig.TARGET_SDK
         versionCode = DefaultConfig.VERSION_CODE
         versionName = DefaultConfig.VERSION_NAME
-
         testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
-        buildConfigField("String", "BASE_URL", properties["BASE_URL"] as String)
-        buildConfigField("String", "USER_ID", properties["USER_ID"] as String)
-        buildConfigField("String", "TEST_ACCESS_TOKEN", properties["TEST_ACCESS_TOKEN"] as String)
-        buildConfigField("String", "TEST_REFRESH_TOKEN", properties["TEST_REFRESH_TOKEN"] as String)
-        buildConfigField("String","KAKAO_NATIVE_KEY", properties["KAKAO_NATIVE_KEY"] as String)
-
+        buildConfigField("String", "BASE_URL", properties.getProperty("BASE_URL"))
     }
 
     buildTypes {
@@ -40,11 +33,8 @@ android {
                 "proguard-rules.pro"
             )
         }
-
         getByName("debug") {
             buildConfigField("String", "BASE_URL", properties["BASE_URL"] as String)
-            buildConfigField("String", "USER_ID", properties["USER_ID"] as String)
-            buildConfigField("String","KAKAO_NATIVE_KEY", properties["KAKAO_NATIVE_KEY"] as String)
         }
     }
     compileOptions {
