+++ core_dependent/build.gradle.kts
@@ -1,14 +1,9 @@
-import org.jetbrains.kotlin.konan.properties.Properties
-
 plugins {
     id("com.android.library")
     id("org.jetbrains.kotlin.android")
     id("kotlin-parcelize")
     id("kotlin-kapt")
 }
-
-val properties = Properties()
-properties.load(project.rootProject.file("local.properties").inputStream())
 
 android {
     compileSdk = DefaultConfig.COMPILE_SDK
@@ -19,9 +14,6 @@ android {
 
         testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
         consumerProguardFiles("consumer-rules.pro")
-
-        buildConfigField("String", "TEST_ACCESS_TOKEN", properties["TEST_ACCESS_TOKEN"] as String)
-        buildConfigField("String", "TEST_REFRESH_TOKEN", properties["TEST_REFRESH_TOKEN"] as String)
     }
 
     buildTypes {
